kin-data-pipeline/requirements.txt
asn1crypto==1.5.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
idna==3.3
oscrypto==1.3.0
pycparser==2.21
pycryptodomex==3.15.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pytz==2022.1
requests==2.28.0
urllib3==1.26.9
snowflake-connector-python==2.7.9
sqlalchemy==1.4.39
pycoingecko==2.2.0
